+++ src/main/java/com/github/alexdlaird/ngrok/process/NgrokLog.java
@@ -24,10 +24,14 @@
 package com.github.alexdlaird.ngrok.process;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
 import static com.github.alexdlaird.util.StringUtils.isBlank;
+import static java.util.logging.Level.INFO;
+import static java.util.logging.Level.SEVERE;
+import static java.util.logging.Level.WARNING;
 
 /**
  * An object containing a parsed log from the <code>ngrok</code> process.
@@ -36,7 +40,7 @@ public class NgrokLog extends HashMap<String, String> {
 
     private final String line;
     private String t;
-    private String lvl = "INFO";
+    private String lvl = INFO.getName();
     private String msg;
     private String err;
     private String addr;
@@ -67,10 +71,10 @@ public class NgrokLog extends HashMap<String, String> {
                     case "CRIT":
                     case "ERR":
                     case "EROR":
-                        value = "SEVERE";
+                        value = SEVERE.getName();
                         break;
                     case "WARN":
-                        value = "WARNING";
+                        value = WARNING.getName();
                         break;
                 }
             }
@@ -156,6 +160,6 @@ public class NgrokLog extends HashMap<String, String> {
             tokens.add(current.toString());
         }
 
-        return tokens;
+        return Collections.unmodifiableList(tokens);
     }
 }