+++ src/test/java/com/github/alexdlaird/ngrok/process/NgrokProcessTest.java
@@ -5,17 +5,23 @@ import com.github.alexdlaird.ngrok.NgrokTestCase;
 import com.github.alexdlaird.ngrok.conf.JavaNgrokConfig;
 import com.github.alexdlaird.ngrok.installer.NgrokInstaller;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Map;
+import java.util.function.Function;
 
 import static com.github.alexdlaird.ngrok.installer.NgrokInstaller.WINDOWS;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.mock;
 
 public class NgrokProcessTest extends NgrokTestCase {
+
     // TODO: testStart()
 
     // TODO: testStop()
@@ -28,14 +34,14 @@ public class NgrokProcessTest extends NgrokTestCase {
         assertTrue(ngrokProcess.isRunning());
         final Path ngrokPath2 = Paths.get(javaNgrokConfig.getNgrokPath().getParent().toString(), "2", NgrokInstaller.getNgrokBin());
         final Path configPath2 = Paths.get(javaNgrokConfig.getConfigPath().getParent().toString(), "config2.yml");
-        final JavaNgrokConfig javaNgrokConfig2 = new JavaNgrokConfig.Builder()
+        final JavaNgrokConfig javaNgrokConfig2 = new JavaNgrokConfig.Builder(javaNgrokConfig)
                 .withNgrokPath(ngrokPath2)
                 .withConfigPath(configPath2)
                 .build();
         ngrokInstaller.installDefaultConfig(javaNgrokConfig2.getConfigPath(), Map.of("web_addr", ngrokProcess.getApiUrl().substring(7)));
 
         // WHEN
-        final NgrokProcess ngrokProcess2 = new NgrokProcess(javaNgrokConfig2, ngrokInstaller);
+        ngrokProcess2 = new NgrokProcess(javaNgrokConfig2, ngrokInstaller);
         final NgrokException exception = assertThrows(NgrokException.class, ngrokProcess2::start);
 
         // THEN
@@ -50,15 +56,88 @@ public class NgrokProcessTest extends NgrokTestCase {
         assertFalse(ngrokProcess2.isRunning());
     }
 
-    // TODO: testExternalKill()
+    @Test
+    public void testExternalKill() throws InterruptedException {
+        // GIVEN
+        ngrokProcess.start();
+        assertTrue(ngrokProcess.isRunning());
 
-    // TODO: testMultipleProcessesDifferentBinaries()
+        // WHEN
+        final ProcessHandle processHandle = ProcessHandle.allProcesses()
+                .filter(p -> p.info().command().get().contains(javaNgrokConfig.getNgrokPath().toString()))
+                .findFirst().get();
+
+        // THEN
+        processHandle.destroy();
+        long timeoutTime = System.currentTimeMillis() + 10 * 1000;
+        while(processHandle.isAlive() && System.currentTimeMillis() < timeoutTime) {
+            Thread.sleep(50);
+        }
+        assertFalse(processHandle.isAlive());
+        assertFalse(ngrokProcess.isRunning());
+
+        // THEN test we can successfully restart the process
+        ngrokProcess.start();
+        assertTrue(ngrokProcess.isRunning());
+    }
+
+    @Test
+    public void testMultipleProcessesDifferentBinaries() {
+        // GIVEN
+        ngrokInstaller.installDefaultConfig(javaNgrokConfig.getConfigPath(), Map.of("web_addr", "localhost:4040"));
+        final Path ngrokPath2 = Paths.get(javaNgrokConfig.getNgrokPath().getParent().toString(), "2", NgrokInstaller.getNgrokBin());
+        final Path configPath2 = Paths.get(javaNgrokConfig.getConfigPath().getParent().toString(), "config2.yml");
+        final JavaNgrokConfig javaNgrokConfig2 = new JavaNgrokConfig.Builder(javaNgrokConfig)
+                .withNgrokPath(ngrokPath2)
+                .withConfigPath(configPath2)
+                .build();
+        ngrokInstaller.installDefaultConfig(javaNgrokConfig2.getConfigPath(), Map.of("web_addr", "localhost:4041"));
+        ngrokProcess2 = new NgrokProcess(javaNgrokConfig2, ngrokInstaller);
+
+        // WHEN
+        ngrokProcess.start();
+        ngrokProcess2.start();
+
+        // THEN
+        assertTrue(ngrokProcess.isRunning());
+        assertEquals("http://localhost:4040", ngrokProcess.getApiUrl());
+        assertTrue(ngrokProcess2.isRunning());
+        assertEquals("http://localhost:4041", ngrokProcess2.getApiUrl());
+    }
 
     // TODO: testMultipleProcessesSameBinaryFails()
 
-    // TODO: testProcessLogs()
+    @Test
+    public void testProcessLogs() {
+        // WHEN
+        ngrokProcess.start();
 
-    // TODO: testLogEventCallbackAndMaxLogs()
+        // THEN
+        for (final NgrokLog log : ngrokProcess.getProcessMonitor().getLogs()) {
+            assertNotNull(log.getT());
+            assertNotNull(log.getLvl());
+            assertNotNull(log.getMsg());
+        }
+    }
+
+    @Test
+    public void testLogEventCallbackAndMaxLogs() throws InterruptedException {
+        // GIVEN
+        final Function<NgrokLog, Void> logEventCallbackMock = mock(Function.class);
+        final JavaNgrokConfig javaNgrokConfig2 = new JavaNgrokConfig.Builder(javaNgrokConfig)
+                .withLogEventCallback(logEventCallbackMock)
+                .withMaxLogs(5)
+                .build();
+        ngrokProcess2 = new NgrokProcess(javaNgrokConfig2, ngrokInstaller);
+
+        // WHEN
+        ngrokProcess2.start();
+        Thread.sleep(1000);
+
+        // THEN
+        assertTrue(Mockito.mockingDetails(logEventCallbackMock).getInvocations().size() > ngrokProcess2.getProcessMonitor().getLogs().size());
+        assertEquals(5, ngrokProcess2.getProcessMonitor().getLogs().size());
+    }
 
     // TODO: testNoMonitorThread()